+++ XTECDigital/src/app/Vistas/Profesor/ver-evaluaciones/ver-evaluaciones.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute, ParamMap} from '@angular/router';
-import { InfoGrupoService } from 'src/app/Vistas/Profesor/ServiciosProfesor/info-grupo.service';
 import { EvaluacionesService } from 'src/app/Vistas/Profesor/ServiciosProfesor/evaluaciones.service';
 import { Evaluacion } from '../ModelosProfesor/evaluacion';
+import { Estado } from 'src/app/modelos/estado';
+import Swal from 'sweetalert2'
 
 
 @Component({
@@ -21,9 +22,14 @@ export class VerEvaluacionesComponent implements OnInit {
   //Lista de evaluaciones correspondientes a un rubro
   listaEvaluaciones: Evaluacion[] = [];
 
-  constructor(private evaluacionesService: EvaluacionesService, private infoGrupo: InfoGrupoService, private route: ActivatedRoute, private router: Router) { }
+  //Estado actual de la aplicaión
+  estadoLocal: Estado;
+
+  constructor(private evaluacionesService: EvaluacionesService, private route: ActivatedRoute, private router: Router) { }
 
   ngOnInit(): void {
+    //Se carga la información de la aplicación almacenada en local storage
+    this.estadoLocal = JSON.parse(localStorage.getItem('EstadoActual'));
     //primero se guarda el número de cédula del profesor
     this.route.params.forEach((urlParams) => {
       this.nombreRubro= urlParams['nombreRubro'];
@@ -35,27 +41,27 @@ export class VerEvaluacionesComponent implements OnInit {
   }
 
   cerrar(){
-    this.router.navigate(['/ProfesorGrupo',this.infoGrupo.numeroCedula, this.infoGrupo.nombreGrupo, 'Evaluaciones']);
+    this.router.navigate(['/ProfesorGrupo',this.estadoLocal.numeroCedula, this.estadoLocal.nombreProfesor, this.estadoLocal.nombreGrupo, 'Evaluaciones']);
 
   }
 
   nuevaEvaluacion(){
     //Calcular el porcentaje restante que le queda al rubro
-    this.router.navigate(['/ProfesorGrupo',this.infoGrupo.numeroCedula, this.infoGrupo.nombreGrupo, 'NuevaEvaluacion', this.nombreRubro, this.porcentajeRubro ]);
+    this.router.navigate(['/ProfesorGrupo',this.estadoLocal.numeroCedula, this.estadoLocal.nombreProfesor, this.estadoLocal.nombreGrupo, 'NuevaEvaluacion', this.nombreRubro, this.porcentajeRubro ]);
   }
 
   calificarEntregables(evaluacion: Evaluacion){
-    this.router.navigate(['/ProfesorGrupo',this.infoGrupo.numeroCedula, this.infoGrupo.nombreGrupo, 'Entregables', evaluacion.nombre, this.nombreRubro ]);
+    this.router.navigate(['/ProfesorGrupo',this.estadoLocal.numeroCedula, this.estadoLocal.nombreProfesor, this.estadoLocal.nombreGrupo, 'Entregables', evaluacion.nombre, this.nombreRubro ]);
   }
 
   actualizarEvaluaciones(){
     //Se solicitan los datos por medio del servicio de evaluaciones
     this.evaluacionesService.getEvaluaciones(
-      this.infoGrupo.codigoCurso,
+      this.estadoLocal.codigoCurso,
       this.nombreRubro,
-      this.infoGrupo.numeroGrupo,
-      this.infoGrupo.anio,
-      this.infoGrupo.periodo
+      this.estadoLocal.numeroGrupo,
+      this.estadoLocal.anio,
+      this.estadoLocal.periodo
     ).subscribe(
       data => {
         //Se limpia la lista de evaluaciones
@@ -79,23 +85,35 @@ export class VerEvaluacionesComponent implements OnInit {
   }
 
   eliminarEvaluacion(nombreEvaluacion: string){
-    //Se solicita eliminar una evaluacion mediante el servicio de evaluaciones
-    this.evaluacionesService.eliminarEvaluacion(
-      this.infoGrupo.codigoCurso,
-      this.nombreRubro,
-      nombreEvaluacion,
-      this.infoGrupo.numeroGrupo,
-      this.infoGrupo.anio,
-      this.infoGrupo.periodo
-    ).subscribe(
-      data => {
-        console.log(data);
-      },
-      error => {
-      console.log(error);
-      this.actualizarEvaluaciones();
-      });
-
+    Swal.fire({
+      title: 'Eliminar Evaluación',
+      text: "¿ Seguro que deseas cerrar la evaluación " + nombreEvaluacion + " ?",
+      icon: 'warning',
+      showCancelButton: true,
+      cancelButtonColor: '#3085d6',
+      confirmButtonColor: '#d33',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Confirmar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        //Se solicita eliminar una evaluacion mediante el servicio de evaluaciones
+        this.evaluacionesService.eliminarEvaluacion(
+          this.estadoLocal.codigoCurso,
+          this.nombreRubro,
+          nombreEvaluacion,
+          this.estadoLocal.numeroGrupo,
+          this.estadoLocal.anio,
+          this.estadoLocal.periodo
+        ).subscribe(
+          data => {
+            console.log(data);
+          },
+          error => {
+          console.log(error);
+          this.actualizarEvaluaciones();
+          });
+          }
+        })
   }
 
 }
